requirements.txt
streamlit>=1.36.0
python-pptx>=0.6.23
Pillow>=10.0.0
